refactor(engine)!: move credential runtime orchestration
Made-with: Cursor
